Add create methods for Results and refactor
Add Create method for MaterialResult and ElementResult and refactor the evaluate methods and Breakdown methods to make use of these.

Moving the ResultConstructor methods into these creates and making them private
